generate-report/requirements.txt
python-dotenv==1.1.0
openai==1.72.0
pandas==2.2.3
python-docx==1.1.2
matplotlib==3.10.7
reportlab==4.4.5